# fix: better performance for labelled matrix display