Current master:
```
julia> t = character_table("B");  # if you really want to see the formatted output then omit the semicolon

julia> @time show(IOBuffer(), "text/plain", t);
  9.962639 seconds (74.39 M allocations: 4.077 GiB, 5.35% gc time, 76.55% compilation time)

julia> @time show(IOBuffer(), "text/plain", t);
  2.275388 seconds (20.78 M allocations: 1.275 GiB, 3.93% gc time)
```
With this PR:
```
julia> t = character_table("B");  # if you really want to see the formatted output then omit the semicolon

julia> @time show(IOBuffer(), "text/plain", t);
  2.590985 seconds (8.90 M allocations: 575.139 MiB, 10.43% gc time, 95.47% compilation time)

julia> @time show(IOBuffer(), "text/plain", t);
  0.130060 seconds (1.57 M allocations: 76.671 MiB)
```